backend/requirements.txt
fastapi
uvicorn[standard]
httpx
openai
python-dotenv
psycopg2-binary
supabase
sqlalchemy
pydantic
beautifulsoup4
tqdm
python-multipart
pgvector
